nse_monitor/requirements.txt
requests
pymupdf
python-dotenv
google-generativeai
google-genai
pytest
